map arrays deep too
